refactor(platform): rename `DANSWER_RUNNING_IN_DOCKER`->`ONYX_RUNNING_IN_DOCKER`
